Add damage calcs
